modified dtos for FE

+++ src/main/java/nhom8/javabackend/hotel/hotel/dto/FindHotelByMostBookingDto.java
@@ -28,7 +28,7 @@ public interface FindHotelByMostBookingDto {
 
 	public String getCondition();
 
-	public int getRating();
+	public float getRating();
 
 	public int getRatingCount();
